Make authz lookup more transfer efficient (filter at DB) (#1784)
The GetValidAuthorizations database query does not filter out the most recent
authorization before transmitting the query results back to the SA. This
changes that query to a "max-value-in-group"-type query that filters at the
database side.

If any authorizations for a given identifier have the exact same expiration
date, this query will return duplicate entries. The existing code added in
PR #1784 will effectively deduplicate, so this should be a drop-in replacement
to speed up the query.

An EXPLAIN on this query shows the derived query as "Using where, using index."